requirements.txt
Django==1.5.5
django-admin-tools==0.5.2
django-ckeditor==4.0.2
django-guardian==1.2.5
django-modeltranslation==0.8
django-tinymce==2.7.0
django-userena==1.4.0
easy-thumbnails==2.2
html2text==2014.12.29
Pillow==2.7.0
psycopg2==2.6
six==1.9.0
South==1.0.2
